+++ src/onegov.org/onegov/org/views/form_collection.py
@@ -1,21 +1,22 @@
 """ Lists the custom forms. """
 
 from onegov.core.security import Public
-from onegov.form import FormCollection
+from onegov.form import FormCollection, FormDefinition
 from onegov.org import _, OrgApp
 from onegov.org.layout import FormCollectionLayout
-from unidecode import unidecode
+from onegov.org.utils import group_by_column
 
 
 @OrgApp.html(model=FormCollection, template='forms.pt', permission=Public)
 def view_form_collection(self, request):
 
-    # XXX add collation support to the core (create collations automatically)
-    forms = self.definitions.query().all()
-    forms = sorted(forms, key=lambda d: unidecode(d.title))
-
     return {
         'layout': FormCollectionLayout(self, request),
         'title': _("Forms"),
-        'forms': request.exclude_invisible(forms)
+        'forms': group_by_column(
+            request=request,
+            query=self.definitions.query(),
+            group_column=FormDefinition.group,
+            sort_column=FormDefinition.order
+        )
     }